requirements.txt
asgiref==3.2.7
autopep8==1.5.1
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.9
mccabe==0.6.1
psycopg2==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
